+++ packages/asc-ui/src/components/Radio/RadioContext.ts
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const RadioContext = React.createContext({
+  checked: '',
+  setChecked: () => {},
+  name: undefined,
+  error: false,
+} as {
+  checked: string
+  setChecked: Function
+  name: string | undefined
+  error: boolean | undefined
+})
+export default RadioContext