+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Prefix the change with one of these keywords:
 
 ## Unreleased
 
+- Added: RadioGroup component.
+- Updated: Radio component. (completely redeveloped)
 - Fixed: IE11 support, by adding `babel-loader` to webpack config
 
 ## Canary